+++ Editor/MergeSkinnedMeshEditor.cs
@@ -30,9 +30,13 @@ namespace Anatawa12.AvatarOptimizer
         {
             ComponentValidation.RegisterValidator<MergeSkinnedMesh>(component =>
             {
-                var err = new List<ErrorLog>();
-                if (component.GetComponent<SkinnedMeshRenderer>().sharedMesh)
-                    err.Add(ErrorLog.Warning("MergeSkinnedMesh:warning:MeshIsNotNone"));
+                var err = new ErrorLog[2];
+                var smr = component.GetComponent<SkinnedMeshRenderer>();
+                if (smr.sharedMesh)
+                    err[0] = ErrorLog.Warning("MergeSkinnedMesh:warning:MeshIsNotNone");
+
+                if (component.renderersSet.GetAsSet().Contains(smr))
+                    err[1] = ErrorLog.Validation("MergeSkinnedMesh:validation:self-recursive");
 
                 return err;
             });