+++ Editor/Processors/SkinnedMeshes/EditSkinnedMeshProcessor.cs
@@ -22,7 +22,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
             Target = component.GetComponent<SkinnedMeshRenderer>();
         }
 
-        public abstract void Process(OptimizerSession session, MeshInfo2 target, MeshInfo2Holder meshInfo2Holder);
+        public abstract void Process(OptimizerSession session, MeshInfo2 target);
 
         public abstract IMeshInfoComputer GetComputer(IMeshInfoComputer upstream);
 
@@ -42,7 +42,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
         IEnumerable<SkinnedMeshRenderer> Dependencies { get; }
         SkinnedMeshRenderer Target { get; }
         EditSkinnedMeshComponent Component { get; }
-        void Process(OptimizerSession session, MeshInfo2 target, MeshInfo2Holder meshInfo2Holder);
+        void Process(OptimizerSession session, MeshInfo2 target);
 
         [NotNull] IMeshInfoComputer GetComputer([NotNull] IMeshInfoComputer upstream);
     }