+++ Editor/Processors/SkinnedMeshes/FreezeBlendShapeProcessor.cs
@@ -13,7 +13,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.SkinnedMeshes
 
         public override EditSkinnedMeshProcessorOrder ProcessOrder => EditSkinnedMeshProcessorOrder.AfterRemoveMesh;
 
-        public override void Process(OptimizerSession session, MeshInfo2 target, MeshInfo2Holder meshInfo2Holder)
+        public override void Process(OptimizerSession session, MeshInfo2 target)
         {
             FreezeBlendShapes(Target, session, target, Component.FreezingShapeKeys);
         }