+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog].
 
 ## [Unreleased]
 ### Added
+- Pre-building validation for MergeBone `#417`
+  - There are some (rare) cases that are not supported by MergeBone. This adds warning for such case.
+- Validation error for self recursive MergeSkinnedMesh `#418`
 - Advanced Settings Section for Trace and Optimize `#419`
   - Moved `Use Advanced Animator Parser` to there
   - Added `Exclusions` for exclude some GameObjects from optimization
@@ -16,12 +19,18 @@ The format is based on [Keep a Changelog].
 - Improved 'Remove Unused Objects' `#401`
   - Remove Unused Objects now removes unnecessary Components & Bones!
   - You may use `Use Legacy GC` to use legacy algotythm for Remove Unused Objects in `Advanced Settings` (`#419`)
+- Performance: Share MeshInfo2 between SkinnedMesh processing and MergeBone `#421`
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+- Crash with Unity 2022 `#423`
+  - [Due to bug in Unity Editor 2022.3 or later][unity-bug], Avatar Optimizer was not compatible with Unity 2022.
+
+[unity-bug]: https://issuetracker.unity3d.com/issues/crash-on-gettargetassemblybyscriptpath-when-a-po-file-in-the-packages-directory-is-not-under-an-assembly-definition
+
 
 ### Security